BacklogTaskViewer/pyproject.toml
[project]
name = "backlog-task-viewer"
version = "0.1.0"
description = "複数Backlogプロジェクトの個人タスクを横断管理するデスクトップアプリケーション"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
    "PySide6>=6.8.0",
    "PySide6-Addons>=6.8.0",
    "requests>=2.31.0",  # Backlog API v2直接呼び出し用
    "urllib3>=1.26,<2.0",  # requests依存、1.26.x系を明示
    "python-dotenv>=1.0.0",
    "keyring>=24.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "black>=23.7.0",
    "ruff>=0.0.282",
    "mypy>=1.4.1",
    "pyinstaller>=5.13.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
